% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evpi_estimators.R
\name{estimate_evpi_evppi_diff_threshold}
\alias{estimate_evpi_evppi_diff_threshold}
\title{Function to estimate the expected value of partial perfect information
This is the outer loop of the two stage Monte Carlo process}
\source{
https://www.sciencedirect.com/science/article/pii/S1098301510605888
}
\usage{
estimate_evpi_evppi_diff_threshold(
  parameter_of_interest,
  param_file,
  colnames_paramdistr,
  list_markov,
  threshold_values,
  outer_iterations = NULL,
  inner_iterations = NULL,
  comparator = NULL
)
}
\arguments{
\item{parameter_of_interest}{parameter of interest}

\item{param_file}{all parameters required to run the model,provided
with name of the parameter, distribution and parameters that define
the probability distribution}

\item{colnames_paramdistr}{column names where parameter distribution and
values are defined}

\item{list_markov}{list of markov models to estimate the NMB}

\item{threshold_values}{threshold values of WTP}

\item{outer_iterations}{number of iterations for outer loop}

\item{inner_iterations}{number of iterations for inner loop}

\item{comparator}{optional parameter if need ICER as the output}
}
\value{
result result for evpi and evppi for different values of threshold
}
\description{
Function to estimate the expected value of partial perfect information
This is the outer loop of the two stage Monte Carlo process
}
