% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evpi_estimators.R
\name{estimate_evpi_evppi_single_param}
\alias{estimate_evpi_evppi_single_param}
\title{Function to estimate the expected value of partial perfect information
This is the outer loop of the two stage Monte Carlo process}
\source{
https://www.sciencedirect.com/science/article/pii/S1098301510605888
}
\usage{
estimate_evpi_evppi_single_param(
  param_interest,
  param_file,
  colnames_paramdistr,
  list_markov,
  threshold,
  outer_iterations,
  inner_iterations,
  comparator = NULL
)
}
\arguments{
\item{param_interest}{the main parameter of interest}

\item{param_file}{all parameters required to run the model,provided with name of
parameter, distribution and parameters that define the probability distribution}

\item{colnames_paramdistr}{coumn names where parameter distribution and values are defined}

\item{list_markov}{list of markov models to estimate the NMB}

\item{threshold}{threshold values of WTP}

\item{outer_iterations}{number of iterations for outer loop}

\item{inner_iterations}{number of iterations for inner loop}

\item{comparator}{optinal parameter if need ICER as the output}
}
\value{
result evpi and evppi for a single param
}
\description{
Function to estimate the expected value of partial perfect information
This is the outer loop of the two stage Monte Carlo process
}
