% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evpi_estimators.R
\name{find_nmb_allparams}
\alias{find_nmb_allparams}
\title{Function to get NMB for a set of parameter
This is the inner loop of the two stage Monte Carlo process}
\usage{
find_nmb_allparams(
  param_file,
  colnames_paramdistr,
  list_markov,
  threshold,
  comparator = NULL
)
}
\arguments{
\item{param_file}{all parameters required to run the model,provided with
name of parameter, distribution and parameters that define the probability
distribution}

\item{colnames_paramdistr}{column names where parameter distribution and
values are defined}

\item{list_markov}{list of markov models to estimate the NMB}

\item{threshold}{threshold values of WTP}

\item{comparator}{optinal parameter if need ICER as the output}
}
\value{
icer_nmb ICER and NMB using all sampled parameters
}
\description{
Function to get NMB for a set of parameter
This is the inner loop of the two stage Monte Carlo process
}
\details{
For the given set of parameter find the nmb from the packDAMipd function
calculate_icer_nmb
}
\keyword{internal}
