% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evpi_estimators.R
\name{get_all_parameter_list}
\alias{get_all_parameter_list}
\title{Function to estimate the expected value of partial perfect information}
\usage{
get_all_parameter_list(
  names_params_needed,
  names_params_model,
  params_passed,
  param_file,
  colnames_paramdistr
)
}
\arguments{
\item{names_params_needed}{names of needed parameters  from param matrix returned}

\item{names_params_model}{names of parameters in the model}

\item{params_passed}{parameters passed while running the markov model}

\item{param_file}{all parameters required to run the model,provided with name of
parameter, distribution and parameters that define the probability distribution}

\item{colnames_paramdistr}{col names where the parameter distribution is defined}
}
\value{
current_parameter_list set of all parameters
}
\description{
Function to estimate the expected value of partial perfect information
}
\details{
this function gets all the parameters except the parameter of interest
if they parameter is fixed, just read from file or a distribution,
then gets it from a distribution, or if it to be calculated, just
give it back as it is
}
\keyword{internal}
