% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evpi_estimators.R
\name{plot_evpi_threshold}
\alias{plot_evpi_threshold}
\title{Function to estimate the expected value of partial perfect information
This is the outer loop of the two stage Monte Carlo process}
\usage{
plot_evpi_threshold(result_evpi_evppi)
}
\arguments{
\item{result_evpi_evppi}{result from estimation of evpi and evppi}
}
\value{
plot the plot
}
\description{
Function to estimate the expected value of partial perfect information
This is the outer loop of the two stage Monte Carlo process
}
